# server/docs/swagger.json
{
    "produces": [
        "application/json"
    ],
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "AgentGame server API for game management, NPCs, players, combat, and more.",
        "title": "AgentGame API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
...
        "/scenes": {
            "get": {
                "description": "Get paginated list of scenes",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "scenes"
                ],
                "summary": "List scenes",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "Page size",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new scene",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "scenes"
                ],
                "summary": "Create a scene",
                "parameters": [
                    {
                        "description": "Scene data",
                        "name": "scene",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_ddc-111_agentGame_server_internal_database_models.Scene"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/scenes/{id}": {
            "get": {
                "description": "Get a scene by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "scenes"
                ],
                "summary": "Get a scene",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Scene ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "put": {
                "description": "Update a scene by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "scenes"
                ],
                "summary": "Update a scene",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Scene ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Scene data",
                        "name": "scene",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_ddc-111_agentGame_server_internal_database_models.Scene"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete a scene by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "scenes"
                ],
                "summary": "Delete a scene",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Scene ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "github_com_ddc-111_agentGame_server_internal_database_models.Agent": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "deleted_at": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "knowledge_base": {
                    "description": "JSON数组",
                    "type": "string"
                },
                "llm_model": {
                    "type": "string"
                },
                "llm_provider": {
                    "type": "string"
                },
                "max_messages": {
                    "type": "integer"
                },
                "max_tokens": {
                    "type": "integer"
                },
                "memory_type": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "summary_enabled": {
                    "type": "boolean"
                },
                "system_prompt": {
                    "type": "string"
                },
                "temperature": {
                    "type": "number"
                },
                "tools": {
                    "description": "JSON数组",
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "github_com_ddc-111_agentGame_server_internal_database_models.Item": {
            "type": "object",
            "properties": {
                "category": {
                    "type": "string"
                },
                "code": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "deleted_at": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "description": {
                    "type": "string"
                },
                "effect": {
                    "description": "JSON对象",
                    "type": "string"
                },
                "icon": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "github_com_ddc-111_agentGame_server_internal_database_models.NPC": {
            "type": "object",
            "properties": {
                "agent": {
                    "$ref": "#/definitions/github_com_ddc-111_agentGame_server_internal_database_models.Agent"
                },
                "agent_id": {
                    "type": "integer"
                },
                "avatar": {
                    "type": "string"
                },
                "behaviors": {
                    "description": "JSON数组",
                    "type": "string"
                },
                "code": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "deleted_at": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "schedule": {
                    "description": "JSON数组",
                    "type": "string"
                },
                "shop": {
                    "$ref": "#/definitions/github_com_ddc-111_agentGame_server_internal_database_models.Shop"
                },
                "shop_id": {
                    "type": "integer"
                },
                "sprite": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "github_com_ddc-111_agentGame_server_internal_database_models.Portal": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "deleted_at": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "id": {
                    "type": "integer"
                },
                "scene_id": {
                    "type": "integer"
                },
                "target_scene": {
                    "type": "string"
                },
                "target_x": {
                    "type": "integer"
                },
                "target_y": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                },
                "x": {
                    "type": "integer"
                },
                "y": {
                    "type": "integer"
                }
            }
        },
...
        "github_com_ddc-111_agentGame_server_internal_database_models.Scene": {
            "type": "object",
            "properties": {
                "background": {
                    "type": "string"
                },
                "code": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "deleted_at": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "description": {
                    "type": "string"
                },
                "height": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "portals": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_ddc-111_agentGame_server_internal_database_models.Portal"
                    }
                },
                "scene_npcs": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_ddc-111_agentGame_server_internal_database_models.SceneNPC"
                    }
                },
                "updated_at": {
                    "type": "string"
                },
                "width": {
                    "type": "integer"
                }
            }
        },
        "github_com_ddc-111_agentGame_server_internal_database_models.SceneNPC": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "deleted_at": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "id": {
                    "type": "integer"
                },
                "npc": {
                    "$ref": "#/definitions/github_com_ddc-111_agentGame_server_internal_database_models.NPC"
                },
                "npc_id": {
                    "type": "integer"
                },
                "scene": {
                    "$ref": "#/definitions/github_com_ddc-111_agentGame_server_internal_database_models.Scene"
                },
                "scene_id": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                },
                "x": {
                    "type": "integer"
                },
                "y": {
                    "type": "integer"
                }
            }
        },
        "github_com_ddc-111_agentGame_server_internal_database_models.Shop": {
            "type": "object",
            "properties": {
                "close_time": {
                    "type": "string"
                },
                "code": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "deleted_at": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "description": {
                    "type": "string"
                },
                "discount": {
                    "description": "JSON对象",
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_ddc-111_agentGame_server_internal_database_models.ShopItem"
                    }
                },
                "name": {
                    "type": "string"
                },
                "open_time": {
                    "type": "string"
                },
                "owner_npc": {
                    "type": "string"
                },
                "scene_code": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "github_com_ddc-111_agentGame_server_internal_database_models.ShopItem": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "deleted_at": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "id": {
                    "type": "integer"
                },
                "item": {
                    "$ref": "#/definitions/github_com_ddc-111_agentGame_server_internal_database_models.Item"
                },
                "item_id": {
                    "type": "integer"
                },
                "price": {
                    "type": "integer"
                },
                "shop": {
                    "$ref": "#/definitions/github_com_ddc-111_agentGame_server_internal_database_models.Shop"
                },
                "shop_id": {
                    "type": "integer"
                },
                "stock": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "gorm.DeletedAt": {
            "type": "object",
            "properties": {
                "time": {
                    "type": "string"
                },
                "valid": {
                    "description": "Valid is true if Time is not NULL",
                    "type": "boolean"
                }
            }
        }
    }
}